C++: rename graph gen class files
