+++ siteupdate/cplusplus/classes/GraphGeneration/HGEdge.cpp
