+++ siteupdate/cplusplus/classes/GraphGeneration/HGEdge.h
