+++ siteupdate/cplusplus/classes/GraphGeneration/HGVertex.cpp
