+++ siteupdate/cplusplus/siteupdate.cpp
@@ -70,13 +70,13 @@ class HGEdge;
 #include "classes/TravelerList/TravelerList.cpp"
 #include "classes/HighwaySegment/HighwaySegment.cpp"
 #include "classes/GraphGeneration/HighwayGraphEdgeInfo.h"
-#include "classes/GraphGeneration/HighwayGraphCollapsedEdgeInfo.h"
-#include "classes/GraphGeneration/HighwayGraphVertexInfo.cpp"
+#include "classes/GraphGeneration/HGEdge.h"
+#include "classes/GraphGeneration/HGVertex.cpp"
 #include "classes/GraphGeneration/PlaceRadius.cpp"
 #include "classes/GraphGeneration/GraphListEntry.cpp"
 #include "classes/GraphGeneration/HighwayGraph.cpp"
 #include "classes/GraphGeneration/HighwayGraphEdgeInfo.cpp"
-#include "classes/GraphGeneration/HighwayGraphCollapsedEdgeInfo.cpp"
+#include "classes/GraphGeneration/HGEdge.cpp"
 #include "threads/ReadWptThread.cpp"
 #include "threads/NmpMergedThread.cpp"
 #include "threads/ReadListThread.cpp"
